# api/openapi.yaml
openapi: 3.0.0
info:
  title: Truther API
  description: Truther API Documentation
  version: 1.0.0
components:
  schemas: {}
paths:
  /coins/search:
    get:
      operationId: searchCoins
      summary: Search coins
      tags:
        - Coins
      description: Search coins
      parameters:
        - schema:
            type: string
          in: query
          name: query
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        symbol:
                          type: string
                          nullable: true
                        image:
                          type: string
                          nullable: true
                      required:
                        - id
                        - name
                      additionalProperties: false
                required:
                  - items
                additionalProperties: false
  /coins/{id}/market:
    get:
      operationId: getCoinMarket
      summary: Get coin market
      tags:
        - Coins
      description: Get coin market
      parameters:
        - schema:
            type: string
            enum:
              - usd
              - eur
              - brl
            default: usd
          in: query
          name: currency
          required: false
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  symbol:
                    type: string
                  image:
                    type: string
                  currentPrice:
                    type: number
                  marketCap:
                    type: number
                  marketCapRank:
                    type: number
                  volume:
                    type: number
                  volumeHigh24h:
                    type: number
                  volumeHLow24h:
                    type: number
                required:
                  - id
                  - name
                  - symbol
                  - image
                  - currentPrice
                  - marketCap
                  - marketCapRank
                  - volume
                  - volumeHigh24h
                  - volumeHLow24h
                additionalProperties: false
  /users:
    post:
      operationId: createUser
      summary: Create user
      tags:
        - Users
      description: Create user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum:
                    - customer
                    - admin
                  nullable: true
                  default: customer
              required:
                - name
                - email
                - password
              additionalProperties: false
        required: true
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
                  password:
                    type: string
                  role:
                    type: string
                    enum:
                      - customer
                      - admin
                    nullable: true
                    default: customer
                required:
                  - name
                  - email
                  - password
                additionalProperties: false
    get:
      operationId: listUsers
      summary: List users
      tags:
        - Users
      description: List users
      parameters:
        - schema:
            type: number
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
        - schema:
            type: number
            minimum: 10
            default: 10
          in: query
          name: pageSize
          required: false
        - schema:
            type: string
          in: query
          name: name
          required: false
        - schema:
            type: string
          in: query
          name: email
          required: false
        - schema:
            type: string
            enum:
              - customer
              - admin
          in: query
          name: role
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        email:
                          type: string
                        role:
                          type: string
                          enum:
                            - customer
                            - admin
                          nullable: true
                          default: customer
                      required:
                        - name
                        - email
                      additionalProperties: false
                  pagination:
                    type: object
                    properties:
                      page:
                        type: number
                      pageSize:
                        type: number
                      pageCount:
                        type: number
                      total:
                        type: number
                    required:
                      - page
                      - pageSize
                      - pageCount
                      - total
                    additionalProperties: false
                required:
                  - items
                  - pagination
                additionalProperties: false
  /users/{id}:
    put:
      operationId: updateUser
      summary: Update user
      tags:
        - Users
      description: Update user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum:
                    - customer
                    - admin
                  nullable: true
                  default: customer
              required:
                - name
                - email
                - password
              additionalProperties: false
        required: true
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
                  password:
                    type: string
                  role:
                    type: string
                    enum:
                      - customer
                      - admin
                    nullable: true
                    default: customer
                required:
                  - name
                  - email
                  - password
                additionalProperties: false
servers:
  - url: http://localhost:3333
    description: Development